requirements.txt
python-decouple==3.8
termcolor==2.5.0
